Stabilize critical shell console diagnostics
