+++ tests/support/pageAssertions.ts
@@ -4,20 +4,12 @@ import {
   attachRedactedTextArtifact,
   sanitizeArtifactName,
 } from "./artifactRedaction";
-
-export type PageDiagnostics = {
-  consoleErrors: string[];
-  pageErrors: string[];
-};
-
-const BENIGN_CONSOLE_ERROR_PATTERNS = [
-  /Failed to load resource: the server responded with a status of (403|404)/i,
-  /net::ERR_ABORTED/i,
-];
-
-function isBenignConsoleError(message: string) {
-  return BENIGN_CONSOLE_ERROR_PATTERNS.some((pattern) => pattern.test(message));
-}
+export {
+  assertNoConsoleErrors,
+  assertNoPageErrors,
+  type PageDiagnostics,
+} from "./consoleDiagnostics";
+import type { PageDiagnostics } from "./consoleDiagnostics";
 
 export function attachPageDiagnostics(page: Page): PageDiagnostics {
   const diagnostics: PageDiagnostics = {
@@ -60,30 +52,6 @@ export async function attachPageDiagnosticsArtifact(
       "Console errors:",
       ...diagnostics.consoleErrors,
     ].join("\n")
-  );
-}
-
-export function assertNoPageErrors(diagnostics: PageDiagnostics) {
-  if (diagnostics.pageErrors.length === 0) {
-    return;
-  }
-
-  throw new Error(
-    `Unexpected browser page error(s):\n${diagnostics.pageErrors.join("\n")}`
-  );
-}
-
-export function assertNoConsoleErrors(diagnostics: PageDiagnostics) {
-  const actionable = diagnostics.consoleErrors.filter(
-    (message) => !isBenignConsoleError(message)
-  );
-
-  if (actionable.length === 0) {
-    return;
-  }
-
-  throw new Error(
-    `Unexpected browser console error(s):\n${actionable.join("\n")}`
   );
 }
